% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrachecks.R
\name{extrachecks}
\alias{extrachecks}
\title{Run all Extra Checks}
\usage{
extrachecks(path = ".")
}
\arguments{
\item{path}{default \code{"."}. The path to an R package.}
}
\description{
Run all Extra Checks
}
\details{
This is a bad path https://google.com/query
}
\examples{
\dontrun{
  stop("dont run this code")
}

httr::GET("google.com")
}
